requirements.txt
matplotlib==3.5.0
numpy==1.20.3
opencv_python_headless==4.5.4.60
pandas==1.3.4
Pillow==10.0.1
scikit_learn==1.3.0
tensorboardX==2.6.2.2
timm==0.8.15.dev0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
